requirements.txt
langchain
langgraph
streamlit
pandas
numpy
yfinance
requests
matplotlib
plotly
beautifulsoup4
nltk
textblob
python-dotenv
scikit-learn
ta
pydantic
openai
scipy
flask
flask-cors